Restore tokens in dilemma
Restore misinformation tokens in the dilemma scene since it makes more sense to have them there, although it does affect the game balance a bit.  Usually misinformation tokens are dropped into icons though to remove hats so this gives a good reason not to do that.  Also put the scales back since balance is trying to be conveyed.